+++ mayan/apps/documents/api_views/document_file_api_views.py
@@ -26,7 +26,7 @@ class APIDocumentFileListView(
     get: Return a list of the selected document's files.
     post: Create a new document file.
     """
-    ordering_fields = ('comment', 'encoding', 'id', 'mime_type')
+    ordering_fields = ('comment', 'encoding', 'id', 'mimetype')
     serializer_class = DocumentFileSerializer
 
     def create(self, request, *args, **kwargs):